+++ packages/visual-remix-reporter/package.json
@@ -18,9 +18,9 @@
     "build:report": "remix vite:build",
     "build:scripts": "tsc -p tsconfig.scripts.json",
     "clean": "rimraf coverage build *.tsbuildinfo",
-    "dev": "npm run generate-thumbnails && run-p watch:scripts dev:remix",
+    "dev": "npm run script:prepare.report && run-p watch:scripts dev:remix",
     "dev:remix": "remix vite:dev",
-    "generate-thumbnails": "node ./app/scripts/generateThumbnails.mjs",
+    "script:prepare.report": "node ./app/scripts/prepareReport.mjs",
     "test:lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
     "start": "remix-serve ./build/server/index.js",
     "typecheck": "tsc",
